Tdying up  index.js and mongoose schemas
