+++ Database/Models/FcmRegistrationSchema.js
@@ -1,6 +1,6 @@
 var mongoose = require('mongoose');
 
-module.exports = {
+var fcm = new mongoose.Schema({
             _id : { type: mongoose.Schema.ObjectId, auto: true },
             updated_at: { type: Date, default: Date.now },
             registrationTokenId : String,
@@ -8,4 +8,6 @@ module.exports = {
             lastModified : String,
             channels : [String],
             courses : [String]
-        }
+});
+
+module.exports = mongoose.model('FcmRegistrations', fcm);
