+++ Database/Models/MaterialSchema.js
@@ -1,6 +1,6 @@
 var mongoose = require('mongoose');
 
-module.exports = {
+var material =  new mongoose.Schema({
             _id : { type: mongoose.Schema.ObjectId, auto: true },
             updated_at: { type: Date, default: Date.now },
             url : String,
@@ -9,4 +9,6 @@ module.exports = {
             description : String,
             baseUrl : String,
             imageUrl : String
-}
+});
+
+module.exports = mongoose.model('Materials', material);
