+++ Database/Models/RatingSchema.js
@@ -1,6 +1,6 @@
 var mongoose = require('mongoose');
 
-module.exports = {
+var ratings = new mongoose.Schema({
             _id : { type: mongoose.Schema.ObjectId, auto: true },
             updated_at: { type: Date, default: Date.now },
             userMailAddress : String,
@@ -10,4 +10,6 @@ module.exports = {
             comment : String,
             createdAt : String,
             lastModified : String
-}
+})
+
+module.exports = mongoose.model('Ratings', ratings);
