+++ Database/Models/SettingsSchema.js
@@ -1,9 +1,11 @@
 var mongoose = require('mongoose');
 
-module.exports = {
+var settingsSchema = new mongoose.Schema({
             _id : { type: mongoose.Schema.ObjectId, auto: true },
             updated_at: { type: Date, default: Date.now },
             lastUpdateAt : String,
             minVersion : Number,
             channels: Array
-}
+});
+
+module.exports = mongoose.model('Settings', settingsSchema);
