+++ routes/settings.js
@@ -1,0 +1,36 @@
+'use strict';
+
+var express = require('express');
+var router = express.Router();
+var mongoose = require('mongoose');
+var passport  = require('passport');
+var jwt = require('jsonwebtoken');
+var config = require('../config/main');
+var emitter = require('../config/emitter');
+
+// Loading both passport strategies
+require('../config/passport')(passport);
+require('../config/passportAdmin')(passport);
+require('../config/passportLogin')(passport);
+
+// Reload passport strategies
+emitter.on('secretChange', function() {
+    require('../config/passport')(passport);
+    require('../config/passportAdmin')(passport);
+});
+
+
+// Settings
+var settings = require('../Database/Models/SettingsSchema');
+
+// Get settings
+router.get('/', passport.authenticate('jwt', { session: false }), function(req, res, next) {
+    settings.findOne(function (err, settings) {
+        if (err) return console.error(err);
+        res.setHeader('Content-Type', 'application/json');
+        res.json(settings);
+    });
+});
+
+
+module.exports = router;
